Update 5thQUESTION.sql
Table t_zuzana_polakova_project_sql_secondary_final is made to show GDP, GINI coeficient, population in distinct country in distinct year (period 2006-2018).
As next step is needed to create (by executing) table for difference of GDP in Czech republic (for other countries values of prices and values of wages are not available). By executing this, it creates table t_result_table_GDP_growth_percent. 
In the next additional table t_GDP_comparing_prices_wages is able to see growth or decline of GDP, prices and wages in one particular year.
Then looking for average growth  of GDP, prices and wages to be able to detect significant growth.
Last commit answer the question - if growth of GDP causes growth of average prices or average wages or both of them.